reservations_backend/reservations_api/urls.py
from django.urls import path

from drf_spectacular.views import (
    SpectacularAPIView,
    SpectacularSwaggerView,
    SpectacularRedocView,
)

from rest_framework import routers

from . import views

router = routers.SimpleRouter()
router.register(r"addresses", views.AddressViewSet)
router.register(r"customers", views.CustomerViewSet)
router.register(r"facilities", views.FacilityViewSet)
router.register(r"reservations", views.ReservationViewSet)

urlpatterns = [
    path("schema/", SpectacularAPIView.as_view(), name="schema"),
    path(
        "schema/swagger-ui/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
    path(
        "schema/redoc/",
        SpectacularRedocView.as_view(url_name="schema"),
        name="redoc",
    ),
    path("users/", views.UserList.as_view()),
    path("users/<int:pk>/", views.UserDetail.as_view()),
]

urlpatterns += router.urls
