pyproject.toml
[project]
name = "reservation-fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-rest-framework>=0.1.0",
    "django>=5.1.1",
]
